Declare geometry functions in the feature decoded file
